requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Audio processing
openai-whisper>=20230314
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0

# Image OCR (handwritten notes, blackboard/whiteboard)
easyocr>=1.7.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Text processing
nltk>=3.8.0
spacy>=3.5.0
regex>=2023.0.0

# Semantic retrieval & verification
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
pyarrow>=14.0.0
networkx>=3.0
matplotlib>=3.8.0

# UI
streamlit>=1.28.0

# LangChain / LangGraph (optional tracing & orchestration)
langchain>=0.2.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Development
pytest>=7.4.0
black>=23.0.0
